refactor(nn): moved simulated layer. MAC operator design simulations do not work
